Pass bounds to optimize_acqf in Alebo (#954)
Summary:
Addresses part of https://github.com/facebook/Ax/issues/953. The main issue here is that the unit tests are fully mocking optimize_acqf and so the recent added validation of the bounds didn't trigger any failures. This should be reworked to actually run a couple of optimization steps so things are tested e2e.

Pull Request resolved: https://github.com/facebook/Ax/pull/954

Reviewed By: saitcakmak

Differential Revision: D36303644

Pulled By: Balandat

fbshipit-source-id: 64666e1787c5a15432e06b60e84b559e7c215379