Move num trials to BenchmarkProblem
Summary:
Wrote this to see how messy it would wind up being to restructure the benchmark problem to include the number of trials. This is a pretty heavily requested feature that I initially pushed back on, but for a number of reasons (db changes, D37842701 making the nightly run super slow) I've had a tentative change of heart.

Pros:
- This is closer to how many of the researchers think about problems (ie a Problem is fundamentally "How far can method X progress on this problem in n trials?").
- More does not involve doing "extra" trials on some problems like D37842701 did

Cons:
- Forces us to lock number of sobol trials in a separate place than the number of BO trials
- Makes it confusing what to do with methods with more than two GenerationSteps (none exist right now but could in the future)
- Breaks up the simplicity of the relationship between benchmark constructs and Ax constructs

dme65 please review the number of trials I selected for each method. I tried to be generous overall based on the recommendations you previously made but lmk if any adjustments should be made:

- Ackley: 50
- Branin: 30
- BraninCurrin: 30
- BraninCurrin_30d: 100
- Hartmann6: 50
- Harmann6_30d: 100
- CNN HPO MNIST: 50
- CNN HPO Fashion_MNIST: 50
- Jenatton: 50
- Powell: 50

Reviewed By: lena-kashtelyan

Differential Revision: D37843541

fbshipit-source-id: 87f58e8df21505ca84abf547b7a6e69c0ecbbda5