Move timeout_hours to SchedulerOptions, add to benchmark (#1037)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/1037

As titled. Users may want to abort benchmark early if it is taking too long.

Note: AggregatedBenchmarkResult has been changed to only take into account the number of trials completed for all replications in its aggregation calculations (ex if a problem is specified for 50 iterations but the underlying experiments are only able to finish `[28, 28, 31, 39, 35, 34, 31, 29, 29, 38]` trials respectively, the AggregatedBenchmarkResult will calculate its traces to be len 28 to ensure each index represents the same number of replications).

Reviewed By: bernardbeckerman

Differential Revision: D38081887

fbshipit-source-id: 53605c21cc449e1d50225f6f0bde070871163eea