adapt the plot function to sensitivity measures dictionary with error bars
Summary:
a sensitivity_measures dictionary can be passed to the plot function. The sensivity values can take the form of a float or an array of floats if the standard error is also plotted.
The name of the importance measure can be passed and added to the title of the figure.

Reviewed By: saitcakmak

Differential Revision: D38108473

fbshipit-source-id: 0c9ff71b8c5d209e236ff5df572068a46e31c28a