Allow passing label dict to cross validation plot
Summary:
Allows specifying a label dict to pass along to the cross validation plot to handle really long metric names.

This is consistent with how the label_dict kwarg has been added to other plots like the trade-off plot.

Allows for the label dict to be a partial listing, and modifies only the names that are in the dict while leaving the others as original.

Differential Revision: D38863101

fbshipit-source-id: 9876d7edd0f07f166919eb4218af8ac185a88f35