Fix flaky tests creating extra files (#1084)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/1084

adds a cleanup option to the testcase

the tests modified in this diff (other than `ax/fb/tests/test_tutorials.py`) were creating an `experiment.json` only some of the time.  Rather than spend a lot of time digging into why when I couldn't reproduce, I just added a mechanism to delete the file if it exists on teardown, by listing `FILES_TO_CLEAN` on the testcase. `ax/fb/tests/test_tutorials.py` was just modified to bring it in line with this new way of doing things.

Reviewed By: lena-kashtelyan

Differential Revision: D38920758

fbshipit-source-id: a734329a07ac9583c7eedac8e74a0f353ba05065