Allow passing label dict to cross validation plot (#1078)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/1078

Allows specifying a label dict to pass along to the cross validation plot to handle really long metric names.

This is consistent with how the label_dict kwarg has been added to other plots like the trade-off plot.

Allows for the label dict to be a partial listing, and modifies only the names that are in the dict while leaving the others as original.

Reviewed By: qingfeng10

Differential Revision: D38863101

fbshipit-source-id: 713d95a2823393a0d0615ab7d4704e8e8e9eaf3a