Small change in documentation of 'Total raw effect'
Summary: 'Total raw effect' was previously described as a percentage, but it is actually the raw multiple factor between the largest and smallest mean. I adjusted the docs, which should not impact any code relying on this field. I also made a small change to `_total_raw_effect` that avoids computing the minium twice.

Reviewed By: lena-kashtelyan

Differential Revision: D39030553

fbshipit-source-id: e572988b5514fdcf51c7ac00605587330b1ff7f0