Small change in documentation of 'Total raw effect' (#1101)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/1101

'Total raw effect' was previously described as a percentage, but it is actually the raw multiple factor between the largest and smallest mean. I adjusted the docs, which should not impact any code relying on this field. I also made a small change to `_total_raw_effect` that avoids computing the minium twice.

Reviewed By: lena-kashtelyan

Differential Revision: D39030553

fbshipit-source-id: f38038d478666d0ea0916138158cc4cbd6c4e5f9