Refactor transforms to work on observations (#1104)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/1104

This is a re-do of D38833175 (https://github.com/facebook/Ax/commit/aa952971e84b828584909f2a2f03254b88dd9649), which was reverted due to a test failure in f6 that wasn't caught by sandcastle.

The source of the test failure is the change to apply untransform_observation_features to a partially specified fixed_features. I adjusted most of the transforms to ensure that they worked correctly on partially specified observation features, but had missed OrderedChoiceEncode and IntToFloat. The particular f6 failure was from trying to compute a Pareto frontier on a model that had an integer parameter, which failed because the fixed feature for IntToFloat didn't work correctly.

This diff is exactly what was landed earlier, plus the changeset in P525514492 which fixes the two transforms.

Reviewed By: saitcakmak

Differential Revision: D39150552

fbshipit-source-id: 3439d9309bc23d4052a6298ec86a080aa08b5dc3