Change botorch benchmark problems to take in TestProblem class and kwargs
Summary:
David ran into a bug when trying to run Hartmann3 benchmarks on fbl -- this turned out to be an issue with our serialization of botorch benchmark problems. Previously we were passing in an instantiated TestProblem, extracting its class and module name, saving those strings, and reinstantiating at load time. If a problem had args passed in (ex. `dim=3` in David's case) they were lost.

Now we pass in a class and kwargs so we always have those to instantiate with whenever we need, like we do for other botorch components.

Reviewed By: dme65

Differential Revision: D39706491

fbshipit-source-id: 50a0fc2d9a77db097caf89b6929db3209b1f5514