Store detailed generator runs on trials (#1160)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/1160

By detailed generator runs, I mean that they have opt config and search space.  Because candidates are often generated using multiple opt configs, there may be multiple GRs that get attached to the trial with the same model key.  However, we do not want this reflected in plots, so we still have the option to create GRs without opt config and search space for plotting.  None of this is exposed in the top level API.

Reviewed By: lena-kashtelyan

Differential Revision: D39712619

fbshipit-source-id: 4ad0511421489970cab606c4b3cf5b34d0f9fa3d