Implement Interactive Loop with candidate pregeneration (#1144)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/1144

In this diff we create an interactive loop API (based loosely off the existing Loop API) in order to facilitate candidate pregeneration. This can be useful in many contexts, espcially interactive experiments in which trial evaluation entails eliciting a response from a real human. Candidate pregeneration uses the time waiting for trail evaluation to generate new candidates from the data available. Note that this is a tradeoff -- a larger lookahead will result in more trials being generated with less data compared to a smaller lookahead (or no lookahead as in conventional Ax usage) and should only be used in contexts where it is necessary for the user to not experience any "lag" while candidates are being generated.

Reviewed By: lena-kashtelyan

Differential Revision: D39505104

fbshipit-source-id: 0a81ac885b2847f1d97672aa23a445a7f9266b09