Automatically set plotting limits (#1173)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/1173

Our default plotting utilities don't handle outliers well and this makes it very hard to see the region of interest. This diff automatically sets the axes limits for our cv and trace plots using ideas similar to box-and-whisker plots. This will hopefully result in a much better user-experience when there are outliers.

Reviewed By: bernardbeckerman

Differential Revision: D39751755

fbshipit-source-id: af6e581f3641332f31a07fce66461d445b492742