Deprecate `transform_outcomes_and_configs=True` (#1200)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/1200

The `transform_outcomes_and_configs` argument in `get_pareto_frontier_and_configs`  seems to be unneccessary (see dicussion on D40269996 (https://github.com/facebook/Ax/commit/7463cd653d6c19d4d901a7210fb8bb3af1344738)) and is a source of confusion, so we are setting it to be `False` and deprecating the argument. Unfortunately, the default is `True` so this would change the default behavior of the function. Fortunately, it is used few enough times (at least internally) that I can go through each usage of the function.

Reviewed By: lena-kashtelyan

Differential Revision: D40318855

fbshipit-source-id: c25f0ecc5c3182196d7563d4760e1bdc7786db49