Handle refetching when all metrics are not Deltoid3Metrics
Summary:
I noticed this error in the nightly notifications job: P536681688.  This situation was created because Jerry created plain `Metric`s to handle metrics with attribute filters.  We may or may not want to do things this way, *but* we don't want to throw errors in the nightly job!
It still won't correctly fetch that specific metric, and may not correctly notify if the metric is in the opt config, but we'll handle that by fully supporting attribute filters at the metric level.

Differential Revision: D40192042

fbshipit-source-id: 345eb298060259c69255c2dadbc9c18e3739144e