Count ABANDONED trials towards failure rate (#1255)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/1255

The scheduler will kill the optimization if too many trials have failed so the user can take a look at whats going on and make changes if needed. This is especially useful in the new setup we have where metric fetch failures can sometimes result in a trial being marked failed or abandoned.

Since both of these states are bad, we should count both towards the max failure rate so the user can be informed and intervene if necessary.

Reviewed By: bernardbeckerman

Differential Revision: D41227143

fbshipit-source-id: 150408209fc2da0e9e54d13bffb539a52dfae50f