Force refit on final report_results call (#1242)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/1242

Followup on recent diff (D40914601) disabling refitting unless there are newly completed trials. It is possible that we may want to refit the model in other cases as well before reporting, e.g., a trial is EarlyStopped and has some updated data. For this reason, we force a refit once right before scheduler termination.

Reviewed By: lena-kashtelyan

Differential Revision: D40986262

fbshipit-source-id: f03adcc1edb6a9d91a36fcc2a7546bb9b207e88c