Remove optimization_config from winsorization config, obtaining it from modelbridge instead (#1267)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/1267

Before this diff, an unnecessary copy of the optimziation config was being passed through the transform_config into the Winsorize transform. This is unnecessary because the modelbridge, which is passed to the Winsorize transform by default, contains an optimization config. This diff removes the extraneous optimization config and updates the code to point to the one on modelbridge.

Reviewed By: dme65, lena-kashtelyan

Differential Revision: D41344838

fbshipit-source-id: 5cef9e3407c07138f8ee929a3b4f96587f0720c2