Add probit link function to Sobol GP mean sensitivity analysis (#1184)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/1184

Right now the GP mean Sobol sensitivity analysis operates directly on the mean of the GP. With a probit link function, we'd like to be able to apply the analysis to the mean of the distribution of p, not the mean of f. The mean of p can be computed analytically in terms of the mean and variance of f, so this adds the appropriate transformation as a function.

Reviewed By: saitcakmak

Differential Revision: D40066277

fbshipit-source-id: 0eaeb5715c44c078c92525f2ca44d22ccfdbf23d