Use Ax plotting functions instead of Plotly in the Scheduler tutorial (#1321)
Summary:
Figures in the scheduler tutorial are not showing up because Plotly figures don't just automatically render properly with html. Instead we need to use the plotting functionality in Ax written for this.

Currently on the website on v:latest (stable is even more broken):
![image](https://user-images.githubusercontent.com/9137425/207741241-aaecd6a3-80f4-47a6-8d15-fb3cce8cc440.png)

After this fix:
![image](https://user-images.githubusercontent.com/9137425/207898976-2eefb6c5-a563-47da-81b3-e15a4d0c456d.png)

Pull Request resolved: https://github.com/facebook/Ax/pull/1321

Reviewed By: Balandat

Differential Revision: D42052422

Pulled By: esantorella

fbshipit-source-id: 7945c58d27e754c3b2cd1034065d52c26bc6ed39