Update `BestPointMixin.get_trace` to extract observations directly from data
Summary: The reliance on a ModelBridge to convert data into tensors was leading to inconsistencies in the results due to the resulting tensors not being sorted by trial indices, which is needed for `get_trace`. This implements a lightweight helper for extracting the Y-tensor directly from the data that keeps the output sorted by the `trial_index`.

Reviewed By: Balandat

Differential Revision: D42221005

fbshipit-source-id: 29b0eeb2e1f9dd72a53d8c39c4043ff07363f793