Skip model fitting in `TorchModelBridge._fit` if the observations haven't changed (#1344)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/1344

This is a safer way of skipping model fitting that relies on observations directly, rather than checking if any new trials completed (which would risk issues with metrics that are available before completion).

This relies on saving a copy of last set of `observations` on `TorchModelBridge` and checks equality with new observations as a condition for skipping the model fitting.

Reviewed By: Balandat

Differential Revision: D42197036

fbshipit-source-id: 8a99383aa08789fcef02bfdfbbcea5dd5482d2fe