Make it possible to infer noise (#1345)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/1345

We currently set the noise to 0 for all benchmark problems even though we often infer the noise in practice. This diff adds an `infer_noise` flag that we can use to control whether different problems should set the noise to a fixed value or `None`, where the latter will trigger inferring the noise in the model. I'm setting `infer_noise=True` by default for all problems.

Reviewed By: saitcakmak, Balandat

Differential Revision: D39413742

fbshipit-source-id: e50ea9d684411a5da09d06899b1a11dac9aca3b3