Move ListSurrogate functionality into base Surrogate (#1324)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/1324

MBM is moving to a new structure in which the BotorchModel may consist of many Surrogates, each of which may either contain a single botorch model or a ModelList. This structure will allow us to use heterogeneous models for our outcomes, but still preserve the convenience of ModelList where applicable.

This diff is the first step in that change: the porting of ListSurrogate's feature set to the base Surrogate (the next diff will remove ListSurrogate entirely). After this change a user will be able to initialize a Surrogate with (or without) all its options as before, and under the hood at construct time Ax will decide whether to use one botorch Model or a ModelList (all of the user specified Model class). If no Model class was provided to the Surrogate then Model classes will be chosen heuristically on a per-outcome basis.

Importantly, note that this is a slight regression from previous behavior where we could specify nearly every single aspect about each botorch model per outcome (all models in a ModelList will either share the provided settings OR be autoset if none are provided). In the new setup if a user wants to control each model per outcome with this much granularity they will need to provide separate surrogates. We made this decision to improve ergonomics, and because no active use cases specified many different models per outcome that then needed to be wrapped in a ModelList down the line. If this changes we can always revisit this decision.

As far as code changes: the bulk are to Surrogate.construct, which has been broken up to dispatch out onto Surrogate._construct_model and Surrogate._construct_model_list. The former does what construct used to do and the latter does what ListSurrogate.construct does. Conversion to block design has also been moved into Surrogate.construct (from BotorchModel.fit).

Reviewed By: lena-kashtelyan

Differential Revision: D42013742

fbshipit-source-id: bcf564e617dea0ed8a3140051a7a670a1721004e