Remove non-determinism in test_timeout (#1378)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/1378

Because of how strict the timeout was in this case, sometimes we wouldnt make it to the GPEI step which would cause fast_botorch_optimize to be unused (which throws an error and asks you to remove the decorator). Since these replications are so short we don't really need to use the decorator anyway, so I am removing it.

Differential Revision: D42557139

fbshipit-source-id: 445adecca5f061c165ad3055438c47ec881a0661