Derelativize optimization_config instead of data when computing constraints (#1315)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/1315

Applying the [relativize_data function](https://www.internalfb.com/code/fbsource/[d787de44b1b1]/fbcode/ax/utils/stats/statstools.py?lines=300) to the data and then constraining according to the relativized constraint does not behave the same as derelativizing the constraint and then applying it to raw data (for example, the conversion of the constraint bound to percentage points is never performed). This diff fixes this where it appears in best_point.py.

Reviewed By: saitcakmak

Differential Revision: D41978338

fbshipit-source-id: abf555892c307358f912d8e2ac0818483d6c3fa5