Make `metric_names` a required input in `Surrogate.construct`
Summary:
`Surrogate.construct` is only called in `Surrogate.fit`, where `metric_names` is a required input. The usage within `Surrogate.construct` also assumes it to be not-None in many cases, so this update makes sense to me.

These changes also eliminate the test failures in D42687367v1 due to missing metric names in `Surrogate.construct` calls.

Differential Revision: D42699205

fbshipit-source-id: 672151ae964de3a9fbfe9683ea2a31b9154677e1