Make `metric_names` a required input in `Surrogate.construct`
Summary:
`Surrogate.construct` is only called in `Surrogate.fit`, where `metric_names` is a required input. The usage within `Surrogate.construct` also assumes it to be not-None in many cases, so this update makes sense to me.

These changes also eliminate the test failures in D42687367v1 due to missing metric names in `Surrogate.construct` calls.

Differential Revision: D42699205

fbshipit-source-id: 04d0d100aa7395f30b71e0ca120b0365074e4eaf