Make `metric_names` a required input in `Surrogate.construct` (#1387)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/1387

`Surrogate.construct` is only called in `Surrogate.fit`, where `metric_names` is a required input. The usage within `Surrogate.construct` also assumes it to be not-None in many cases, so this update makes sense to me.

These changes also eliminate the test failures in D42687367v1 due to missing metric names in `Surrogate.construct` calls.

Reviewed By: mpolson64

Differential Revision: D42699205

fbshipit-source-id: db7dd08601577d6b37f70d30df0017b24c224cd2