Add hypervolume trace to `get_standard_plots` (#1390)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/1390

Add a new plot for MOO experiments that shows the hypervolume plotted against the iteration num. This is useful for understanding if the pareto frontier is expanding or if the optimization has stalled out.

Two ideas discussed that we ended up not including:
- Use the "magma" color scheme to give the line the same gradient effect as points on the pareto plot (where color corresponds to trial index)
- Normalize the hypervolume to (0, 1) where 1 is the current hypervolume

Open to rethinking these decisions and adding them -- let me know what you think.

Reviewed By: bernardbeckerman

Differential Revision: D42725693

fbshipit-source-id: 64503b663afb70c114f1ce9e2ee6d40361fbbe92