Refactor BestPointMixin to be used as a mixin (#1389)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/1389

The `get_trace` and `get_trace_by_progression` methods were being called directly as static methods when we really want methods on this mixin to be called by the classes that implement the mixin.

For example we used to call like this `BestPointMixin.get_trace(scheduler.experiment)` but what we really want is to be able to call like this `scheduler.get_trace()`.

After this diff each of these methods follow the same structure as the other best point abstract methods like `get_best_point` and `get_hypervolume`

Reviewed By: bernardbeckerman

Differential Revision: D42722280

fbshipit-source-id: 46cb58608014a9840018f0b9c30119213ae24ddb