Use strict rounding if randomized rounding fails in IntToFloat (#1404)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/1404

`IntToFloat` tries to randomly round observations in the inverse transform in order to satisfy the bound and potential parameter constraints. This doesn't play well with `rejection_sample` which may produce candidates that cannot be rounded to a feasible candidate in this way. If `IntToFloat` fails to randomly round the observation it will just return the last rounded observation it tried, which could potentially be outside the search space bounds and cause weird issues with the next transforms, e.g., trying to take the log of 0.

A potential workaround for now is to return a strictly rounded candidate if `IntToFloat` fails. This is guaranteed to satisfy the search space bounds and will play well with the next transforms, but not necessarily the parameter constraints. `rejection_sample` will deal with this properly.

The long-term solution here is likely to revamp the rounding logic in `IntToFloat`, but this should unblock this particular issue for now.

Reviewed By: bernardbeckerman, Balandat

Differential Revision: D42894672

fbshipit-source-id: d2e00cd54e2cd8ebeee1db1db824550d77e6bf2e