Don't try to save lists of python primitives (#1440)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/1440

There was generating a long log warning in N3122283.  In looks like `copy_db_ids()` just traverses down the hierarchy of an object trying to put db ids on everything.

This should be a no op other than raising an error which was _always_ caught and logged.  I think this error was originally meant for catching lists of `Base` where `SortableBase` was needed.

Reviewed By: lena-kashtelyan

Differential Revision: D43475289

fbshipit-source-id: cb597abadcb80d113558296349fc85209a61405f