Extend arm shortname to 8 characters
Summary:
Sterling found a bug where if there are too many (on the order of 16^4) unnamed arms on an experiment, when you call gr.param_df there can be collisions which result in dropped arms in the pandas df.

Frankly I do not know what this shortname is used for (and it looks like we dont use gr.param_df internally) but this is an easy thing to do to fix a corner case. Wdyt?

https://github.com/facebook/Ax/issues/1437

Reviewed By: bernardbeckerman

Differential Revision: D43479351

fbshipit-source-id: 56c8dfee9740a6275618404405fa4f1de1d8c155