Remove MOO_Modular
Summary:
Craig reached out to me because he stumbled into this and was confused whether to use this or BOTORCH_MODULAR for MOO. As far as I can tell this is now just an alais for the same setup as BOTORCH_MODULAR (maybe we consolodated functionality at some point?) and it has no usage that Im aware of -- should we just clean it up to eliminate any confusion?

If this isnt actually dead code please let me know and I will abandon.

Reviewed By: saitcakmak

Differential Revision: D43583980

fbshipit-source-id: e6b0003ad43c977c99be550b6dfece5197b1c397