Remove MOO_Modular (#1459)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/1459

Craig reached out to me because he stumbled into this and was confused whether to use this or BOTORCH_MODULAR for MOO. As far as I can tell this is now just an alais for the same setup as BOTORCH_MODULAR (maybe we consolodated functionality at some point?) and it has no usage that Im aware of -- should we just clean it up to eliminate any confusion?

If this isnt actually dead code please let me know and I will abandon.

Reviewed By: saitcakmak

Differential Revision: D43583980

fbshipit-source-id: 293b709b9c83ea9b5a8afc57bdac5f4bf2f233d7