Feature/json decoder with pathlib path and functions (#1418)
Summary:
## Problem
The current `CORE_DECODER_REGISTRY` and `CORE_ENCODER_REGISTRY` does not include `pathlib.Path` instances which people often store paths as. I am trying to encode and decode some custom Ax Metrics and other subclasses that include me sometime storing paths. A problem with the current encoding and decoding options is that they need to be classes (instead of a function) because of this line in `object_from_json`
``` python
elif issubclass(_class, SerializationMixin):
```
Which doesn't work with `pathlib.Path` since it also only takes position arguments that get splatted/unpacked. It would be easy to to make it accept functions functions by making the check `elif isclass(_class) and issubclass(_class, SerializationMixin):` which happens for other similar checks a dozen or so lines above this line.

Doing so would also allow encoders and decoders that are `classmethods` which are often used as alternative constructors (Something.from_json, Something.to_json, etc.)

## Changes

- Changed `object_from_json` to allow functions  to be in the `decoder_registry`.
  - This was done as described above by changing `elif issubclass(_class, SerializationMixin):` to `elif isclass(_class) and issubclass(_class, SerializationMixin):`
- add pathlib.Path to `CORE_DECODER_REGISTRY` and `CORE_ENCODER_REGISTRY`
-  Adjust typehints around decoder registries from `Dict[str, Type]` to `Dict[str, Callable[[Dict[str, Any]], Any]]`
- Add `pathlib.Path` to `TEST_CASES` in test_json_store.py

Pull Request resolved: https://github.com/facebook/Ax/pull/1418

Reviewed By: mpolson64

Differential Revision: D43483571

Pulled By: esantorella

fbshipit-source-id: 7cc889ef9be02793632752482e7c083ebc449616