Fix Acquisition.get_botorch_objective_and_transform (#1487)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/1487

This does not pass down the acqf class to the `get_botorch_objective_and_transform` helper, and always passes down `objective_thresholds` if they're available. As a result, `get_botorch_objective_and_transform` determines the type of the objective to construct (MOO or otherwise) based on the availability of `objective_thresholds`. This leads to weird downstream errors if objective thresholds are present when a single objective acquisition function is being used. This change should fix that.

Differential Revision: D43803349

fbshipit-source-id: eea6a0b7c286e3098e662355e2db0e4fdf0a7d92