Replace objective_thresholds with botorch_acqf_class in get_botorch_objective_and_transform (#1492)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/1492

This was using the existance of `objective_thresholds` as a proxy for using a multi objective acquisition function. This is not ideal since it could lead to multi-output objectives being used with single-objective acquisition functions. With this change, we directly use the acquisition class to determine the type of the objective needed.

Reviewed By: Balandat

Differential Revision: D43880923

fbshipit-source-id: 2ee0eb22ab3a9535705cd8864f7ad2f7a1038f30