Errs for curve metrics (#1483)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/1483

When we originally implemented the Results in metric fetching we punted on the best thing to do with Curve metrics, which already had some robustness logic built in via returning empty dataframes on failed queries. Unfortunately this empty df technique has caused some downstream issues in the Scheduler (see D43737988) so I've made the decision to implement Errs instead.

Reviewed By: danielrjiang

Differential Revision: D43744785

fbshipit-source-id: 23c220e7e54feca1d681ed44440fffcb20fb424b