Clean up usage of optimizer_options (#1463)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/1463

Prior to this change, the `options` argument of `optimize_acqf` was being parsed inconsistently across various models. Some would look for the `options` among top level arguments in `optimizer_options`, others would look for explicit `options` dict in `optimizer_options`.

The second approach (explicit `options` dict) is consistent with the signature of `optimize_acqf`, and is being used in MBM. This updates the rest of the code base to make it all consistent with each other.

Differential Revision: D43637014

fbshipit-source-id: bed1e3dcccc43eab7154a8f9b35f1eab972033e8