Update the model in ModelSpec.fit if all non-data args are the same
Summary:
Follow up to https://github.com/facebook/Ax/pull/1344. Calling `ModelBridge._fit` instead of constructing a fresh `ModelBridge` will make sure that we can avoid model fitting when it is not necessary.

Also adds a `fit_on_init` kwarg to `ModelBridge.__init__`, which can be used to skip model fitting when a fitted model is not needed.

Differential Revision: D42270109

fbshipit-source-id: 65350c4b5ecc418fbbe24bc01a6f50c35ce81555