Update the model in ModelSpec.fit if all non-data args are the same (#1507)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/1507

Follow up to https://github.com/facebook/Ax/pull/1344. Calling `ModelBridge._fit` instead of constructing a fresh `ModelBridge` will make sure that we can avoid model fitting when it is not necessary.

Also adds a `fit_on_init` kwarg to `ModelBridge.__init__`, which can be used to skip model fitting when a fitted model is not needed.

Reviewed By: Balandat

Differential Revision: D42270109

fbshipit-source-id: c6f8e7a23256ea0461f6a323fb457fcb25f28a18