Refit or update the modelbridge in `ModelSpec.fit` if all non-data args are the same (#1512)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/1512

This is a redo of https://github.com/facebook/Ax/pull/1507, which was reverted due to issues with order of operations when using a status quo: We need to be setting the train data before setting the status quo, and setting the status quo before transforming the data.

In `ModelSpec.fit`, calling `ModelBridge._fit` instead of constructing a fresh `ModelBridge` will make sure that we can avoid model fitting when it is not necessary (the necessary logic was introduced in https://github.com/facebook/Ax/pull/1344).

This also adds a `fit_on_init` kwarg to `ModelBridge.__init__`, which can be used to skip model fitting when a fitted model is not needed.

Reviewed By: Balandat

Differential Revision: D44096998

fbshipit-source-id: 3c508d1676ad963f1596892053cb45d500fdb97b