Fix pareto front tensor converter
Summary: The pareto front tensor converter model had a bug where it didn't work on experiments with only one arm, because it couldn't convert that to a ChoiceParameter. D42478964 fixed that by changing the transforms, but the new set of transforms broke the PEX use case, and also has broken support for some replay experiments. This fixes both and removes the patch for fixing PEX earlier.

Differential Revision: D44174796

fbshipit-source-id: b5a661330e61fdf1b2e7d298df63702318acda6d