Fix pareto front tensor converter (#1526)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/1526

The pareto front tensor converter model had a bug where it didn't work on experiments with only one arm, because it couldn't convert that to a ChoiceParameter. D42478964 fixed that by changing the transforms, but the new set of transforms broke the PEX use case, and also has broken support for some replay experiments. This fixes both and removes the patch for fixing PEX earlier.

Reviewed By: bernardbeckerman

Differential Revision: D44174796

fbshipit-source-id: 8e666cdbd547392ee06acd1f53b36dcea57fd9fb