Add RangeParameter bounds tolerance
Summary:
Due to numerical precision issues, RangeParameter validation can fail. This can be the case even for parameter configs that are generated by Ax itself (this can happen e.g. if the experiment state is serialized and reloaded).

This change adds a numerical tolerance to avoid these issues. This also adds an error in case the parameter range is extremely small and likely to cause numerical issues.

This is a quick fix; we should do a more comprehensive audit of how floating point precision issues may creep up in Ax and implement a more principled fix.

Reviewed By: bernardbeckerman

Differential Revision: D44245817

fbshipit-source-id: 1bc62a7e7e20b18bad4015cad757ca3c4c818eee