Saving SearchSpaceDigest in BotorchModel's fit (#1561)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/1561

This commit stores the `SearchSpaceDigest` object during a `BotorchModel`'s `fit` call, which allows access to the search space parameters after `fit` has completed for e.g. sensitivity analysis. For a modular BoTorch model, the `SearchSpaceDigest` is already stored in the object upon a `fit` call, so this commit merely syncs the behavior of the two model types in this regard.

Reviewed By: bletham

Differential Revision: D44514681

fbshipit-source-id: 14d8504b469144261ff36087792651a5af56a031