Do not overwrite generation step index when adding custom arms to BatchTrial
Summary:
When adding a generator run to a BatchTrial, we update the generation step index to match the generation step index of the generator run. This ensures that we do not mix arms generated from different generation steps in a trial.

However, we support mixing custom arms with arms generated by the generation strategy. When it comes to adding these arms, a "manual" generator run is generated and attached to the trial. In the current setup, this overwrites the generation step index to None, which breaks the `should_move_to_next_step` logic of generation strategy.

This change makes it so that the manual arms do not overwrite the generation step index. This allows the trial to be associated with the generation step that originally created it, which makes sure that we don't get stuck in the Sobol step indefinitely.

Reviewed By: qingfeng10

Differential Revision: D45298272

fbshipit-source-id: afd71b259827525b2d8d7e7849a7f6a1688fe6bf