Set logging level to WARNING for unit tests (#1591)
Summary:
The default Ax logging level is INFO, and Ax prints a lot of INFO logs. This is annoying for unit tests, where we only want to know when something is wrong, and drowns out useful warnings. This sets the logging level one level higher, to WARNING and above.

Pull Request resolved: https://github.com/facebook/Ax/pull/1591

Test Plan: Ran unit tests and made sure that INFO logs were not printed, but WARNING logs were.

Reviewed By: saitcakmak

Differential Revision: D45245652

Pulled By: esantorella

fbshipit-source-id: 34ed52db2b1526651caee47fafc636e4169ee82d