Add `run_metadata_report_keys` property to `Runner` class (#1625)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/1625

Adds a `run_metadata_report_keys` property to the base `Runner` class. This can be used to specify which of the keys used in the `run_metadata` dict returned by a runner's `run()` method are relevant outside the internals of the runner. These can e.g. be utilized in `Scheduler.report_results()` or other places.

Reviewed By: bernardbeckerman

Differential Revision: D46042501

fbshipit-source-id: 971aee1f21e079d766785220fee308138cb3274a