introduce abstract BaseData class
Summary: Intended as a no-op. Creates an abstract class `BaseData` that does not require columns `mean` and `sem`. Creates `Data` class based on `BaseData` that is functionally identical to the prior implementation of `Data`, allowing us to split off functionality to a future `ArtifactData` class that requires column `value`, that can be non-numeric.

Reviewed By: lena-kashtelyan, esantorella

Differential Revision: D46234115

fbshipit-source-id: c06f973c1c6f8cf0fa9b16fa38d2c671ed838e75