introduce abstract BaseData class (#1645)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/1645

Intended as a no-op. Creates an abstract class `BaseData` that does not require columns `mean` and `sem`. Creates `Data` class based on `BaseData` that is functionally identical to the prior implementation of `Data`, allowing us to split off functionality to a future `ArtifactData` class that requires column `value`, that can be non-numeric.

Reviewed By: lena-kashtelyan, esantorella

Differential Revision: D46234115

fbshipit-source-id: 925dacde8a7940dfb56b1ca3f1f69e99a7e063b5