Remvove unused code
Summary: Remove unused code. The plotting code is unused now that we have much easier ways of accessing the data (through the dashboard or just pulling it from Hive). Some of the mathematical code is unused owing to the previous diff in this stack, moving logic to Dataswarm. Even more code can be deleted once I refactor some of the logging.

Differential Revision: D46917803

fbshipit-source-id: 24683726049a73e6ab82b3521c0a5cfd0a04e345