Remvove unused code (#1677)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/1677

Remove unused code. The plotting code is unused now that we have much easier ways of accessing the data (through the dashboard or just pulling it from Hive). Some of the mathematical code is unused owing to the previous diff in this stack, moving logic to Dataswarm. Even more code can be deleted once I refactor some of the logging.

Reviewed By: Balandat

Differential Revision: D46917803

fbshipit-source-id: 913409063cd948adb3c523b852d1c56a4dde51c0