Catch error when trying to mark trial as running after deployment
Summary:
In some situations a trial may be in a status that is incompatible with being
marked as RUNNING (e.g. CANDIDATE) status when processed in the Scheduler here:
https://github.com/facebook/Ax/blob/main/ax/service/scheduler.py#L1477

This may happen e.g. if the deployment in the Runner fails and the trial is
marked as FAILED.

This diff wraps the marking to RUNNNING in a try/except and logs a warning.

An alternative design would be to mark the trials that fail to deploy (e.g.
b/c of infra issues) not as FAILED but as CANDIDATES and pass richer
information to the Scheduler to be able to decide to retry deploying the trial
at a later time. This would require additional work though, and it's also not
clear whether it's the right thing to do (the trial may be stale at the next
deployment attempt).

Reviewed By: bernardbeckerman

Differential Revision: D46959540

fbshipit-source-id: 4cb45ea7a1c49839c78d169e63cc4d4f22d12666